feat: Allow specifying input files from a .txt file
This commit introduces the ability to provide a list of input files for analysis using a  file.

The  CLI is updated to handle  files for the  argument in  and  commands, as well as for  and  in the  command.

The  has been updated to document this new feature with examples.

This commit also includes user-staged changes to , , and adds  as an example.